[CUDA:Bugfix] Fix int32 overflow in im2col packing for large inputs (NaN/garbage output on conv layers with e*l > INT32_MAX)
On layers where (output pixels) * (packed input channels) exceeds
INT32_MAX (e.g. PP-OCRv4 server det, 1280x1920 input, 256ch 9x9 conv:
153600 * 20736 = 3,185,049,600), two int32 multiplications overflow:

1. callIm2ColPack(): size_t maxCount = e * lp  -> int*int overflow
   produced a garbage grid dimension, so cudaLaunchKernel failed with
   cudaErrorInvalidValue. checkKernelErrors is compiled out in release
   builds, so the launch failure was silent and downstream layers
   consumed uninitialized memory (NaN prob maps, scrambled detections).

2. Im2Col_FilterC_Vec4/FilterC kernels: size_t dst_offset =
   eIndex * l_p + ... -> int*int overflow corrupted the im2col buffer
   write addressing, silently producing wrong values even when the
   launch itself succeeded.

Fix: promote the multiplications to size_t. Also fixed the same
pattern in ConvCutlassExecution.cu, DeconvSingleInputExecution.cu and
MultiInputConvExecution.cu.

Verified on RTX A10G (sm_86, CUDA 13.0, MNN 3.6.1):
- PP-OCRv4 server det at 1280x1920: output maxdiff vs CPU drops from
  1.0 (garbage) to 0.06-0.15 (fp16-mix rounding), NaN count 96170 -> 0.
- Overflow boundary confirmed: layer e*lp 2.12e9 passes, 2.34e9 fails
  before the patch, matching the INT32_MAX threshold exactly.

+++ source/backend/cuda/execution/ConvBaseKernel.cu
@@ -58,7 +58,7 @@ __global__ void Im2Col_FilterC(
         size_t sy = oy * sh + ksy * dh- ph;
 
         const int ic_p = icDiv4 * pack;
-        size_t dst_offset = eIndex * l_p + kI * ic + iz;
+        size_t dst_offset = (size_t)eIndex * (size_t)l_p + (size_t)(kI * ic + iz);
 
         if (sx >= 0 && sx < iw) {
             if (sy >=0 && sy < ih) {
@@ -127,7 +127,7 @@ __global__ void Im2Col_FilterC_Vec4(
         const int ic_p = icDiv4 * pack;
 
         const int iz = iz_4 << 2;
-        size_t dst_offset = eIndex * l_p + kI * ic + iz;
+        size_t dst_offset = (size_t)eIndex * (size_t)l_p + (size_t)(kI * ic + iz);
 
         if (sx >= 0 && sx < iw) {
             if (sy >=0 && sy < ih) {
@@ -264,7 +264,7 @@ void callIm2ColPack(const void* input, void* output, const ConvolutionCommon::Im
     const int ih = info->ih;
     const int ic = info->ic;
 
-    size_t maxCount = e * lp;
+    size_t maxCount = (size_t)e * (size_t)lp;
     size_t block_num = runtime->blocks_num(maxCount);
     size_t block_size = runtime->threads_num();
